pyproject.toml
[project]
name = "stts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elevenlabs>=2.8.0",
    "google-genai>=1.26.0",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "speechrecognition>=3.14.3",
]

[dependency-groups]
dev = ["ruff>=0.12.4", "ty>=0.0.1a15"]

[project.scripts]
stt = "stt.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["stt"]

[tool.ruff]
line-length = 80
indent-width = 4

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

docstring-code-format = true

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]

ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
    "D206",  # docstring should be indented with spaces, not tabs
    "D300",  # Use """triple double quotes""".
    "B904",  # raise ... from ...
    "UP007", # use `f"{var}"` instead of `"{var}".format(var=var)`
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pyright]
pythonVersion = "3.13"

typeCheckingMode = "strict"

reportImportCycles = false

reportPrivateUsage = false
reportUnknownMemberType = false
reportUnknownVariableType = false
reportUntypedFunctionDecorator = false
